# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/books/banner": {
            "get": {
                "tags": [
                    "Banner"
                ],
                "summary": "Get platform banner",
                "operationId": "4bdeb1ca3d234bbdd62bebf17b94f821",
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "500": {
                        "description": "An error occurred while get banner."
                    }
                }
            }
        },
        "/api/books/newest": {
            "get": {
                "tags": [
                    "Book"
                ],
                "summary": "Get 6 newest books",
                "operationId": "f7cb2f516f54ceed25d7449a2d60bc81",
                "responses": {
                    "200": {
                        "description": "Get newest book success"
                    },
                    "500": {
                        "description": "Cannot get newest book!"
                    }
                }
            }
        },
        "/api/books/mostReviewed": {
            "get": {
                "tags": [
                    "Book"
                ],
                "summary": "Get 6 most reviewed books",
                "operationId": "8c764bf3e58ec29a0c3e1d5aa7f3fa19",
                "responses": {
                    "200": {
                        "description": "Get most reviewed book success"
                    },
                    "500": {
                        "description": "Cannot get most reviewed book!"
                    }
                }
            }
        },
        "/api/auth/book": {
            "get": {
                "tags": [
                    "Book"
                ],
                "summary": "Get books belonging to the authenticated user",
                "operationId": "62cebb5cdc5658b1cd7b98c8407604c6",
                "responses": {
                    "200": {
                        "description": "Get my book success"
                    },
                    "401": {
                        "description": "Unauthorized: Missing, invalid, or expired token"
                    },
                    "500": {
                        "description": "Server error while retrieving books"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/books/{bookId}": {
            "get": {
                "tags": [
                    "Book"
                ],
                "summary": "Get book detail",
                "operationId": "97ad3f951101d83aca443cece9bc6d96",
                "parameters": [
                    {
                        "name": "bookId",
                        "in": "path",
                        "description": "Id of the book",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get book detail success"
                    },
                    "500": {
                        "description": "Cannot get book detail!"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}